[FIX] 最终提交版本

+++ src/main/java/com/swu/cjyong/main/entity/dto/MemberCount.java
@@ -4,14 +4,14 @@ import lombok.Data;
 
 @Data
 public class MemberCount {
-    private int memberCount = 0;
-    private int account4Count = 0;
-    private int member_school = 0;
-    private int account4_school = 0;
-    private int member_city = 0;
-    private int account4_city = 0;
-    private int member_district = 0;
-    private int account4_district = 0;
+    private long memberCount = 0;
+    private long account4Count = 0;
+    private long member_school = 0;
+    private long account4_school = 0;
+    private long member_city = 0;
+    private long account4_city = 0;
+    private long member_district = 0;
+    private long account4_district = 0;
 
     public void addAll() {
         memberCount = member_city + member_school + member_district;
